// frontend/quickshift/src/backend-api/models/shift-applicant-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ShiftDto } from './shift-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { UserDto } from './user-dto';

/**
 * 
 * @export
 * @interface ShiftApplicantDto
 */
export interface ShiftApplicantDto {
    /**
     * 
     * @type {string}
     * @memberof ShiftApplicantDto
     */
    '_id': string;
    /**
     * 
     * @type {string}
     * @memberof ShiftApplicantDto
     */
    'company': string;
    /**
     * 
     * @type {string}
     * @memberof ShiftApplicantDto
     */
    'shiftId': string;
    /**
     * 
     * @type {string}
     * @memberof ShiftApplicantDto
     */
    'user': string;
    /**
     * 
     * @type {boolean}
     * @memberof ShiftApplicantDto
     */
    'rejected': boolean;
    /**
     * 
     * @type {ShiftDto}
     * @memberof ShiftApplicantDto
     */
    'shift'?: ShiftDto;
    /**
     * 
     * @type {UserDto}
     * @memberof ShiftApplicantDto
     */
    'userData'?: UserDto;
}